backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
google-generativeai
pymongo
mongoengine
passlib[bcrypt] 
pydantic[email]
PyJWT
bcrypt==3.2.0
schedule
pandas
numpy
bs4